+++ __tests__/Sockets.js
@@ -1,1 +1,0 @@
-import TcpSocket from 'react-native-tcp-socket';