+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [3.4.0](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.3.1...v3.4.0) (2020-02-27)
+
+
+### Features
+
+* **Android:** Connect from "cellular" & "ethernet" interfaces ([14bea9b](https://github.com/Rapsssito/react-native-tcp-socket/commit/14bea9b49db0971ac38142dcc91376d7d3752212))
+
+## [3.3.1](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.3.0...v3.3.1) (2020-02-24)
+
+
+### Bug Fixes
+
+* Fixed TypeScript declaration files not being a module ([3e2213a](https://github.com/Rapsssito/react-native-tcp-socket/commit/3e2213a06c6917c8c0696cf31b905cd9e91f52e7))
+
+# [3.3.0](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.2.9...v3.3.0) (2020-02-24)
+
+
+### Features
+
+* Add TypeScript declaration file ([#29](https://github.com/Rapsssito/react-native-tcp-socket/issues/29)) ([99d26b9](https://github.com/Rapsssito/react-native-tcp-socket/commit/99d26b981e66463baa2be174445f0a946f42ab4d))
+
+## [3.2.9](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.2.8...v3.2.9) (2020-02-20)
+
+
+### Bug Fixes
+
+* Fixed an error when sending a Uint8Array from write() ([4d44d3e](https://github.com/Rapsssito/react-native-tcp-socket/commit/4d44d3e06a32d09bda683a55b4dc4629f51ee58d))
+
+## [3.2.8](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.2.7...v3.2.8) (2020-02-18)
+
+
+### Bug Fixes
+
+* Fixed error removing already removed listeners ([7712876](https://github.com/Rapsssito/react-native-tcp-socket/commit/7712876bc6a61ee2efee57161f2a09dc5ff1b0a1))
+
 ## [3.2.7](https://github.com/Rapsssito/react-native-tcp-socket/compare/v3.2.6...v3.2.7) (2020-02-18)